% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfeatureinfor.R
\name{avg_word_len}
\alias{avg_word_len}
\title{Calculate average word length in a string}
\usage{
avg_word_len(text)
}
\arguments{
\item{text}{A character vector for the average word length extraction.}
}
\value{
A numeric vector representing the average word length in the text.
}
\description{
Calculate average word length in a string
}
\examples{
x <- "Here are some words"
avg_word_len(x)
4
}
