% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfeatureinfor.R
\name{count_punc}
\alias{count_punc}
\title{Count punctuations}
\usage{
count_punc(text)
}
\arguments{
\item{text}{A character vector with, at most, one element.}
}
\value{
A vector containing the number of punctuations.
}
\description{
Count punctuations
}
\examples{
x <- "Hello, World!"
count_punc(x)
2
}
