% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfeatureinfor.R
\name{perc_cap_words}
\alias{perc_cap_words}
\title{Percentage of fully capitalised words in a string}
\usage{
perc_cap_words(text)
}
\arguments{
\item{text}{A character vector with length one containing the piece of text to analyze.}
}
\value{
A vector containing the number of fully capitalised words
}
\description{
Percentage of fully capitalised words in a string
}
\examples{
text <- "This is REALLY Awesome!"
perc_cap_words(text)
25
}
